# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  clinic.RqBulkDeleteClinic:
    properties:
      clinic_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - clinic_ids
    type: object
  clinic.RqBulkUpdateClinic:
    properties:
      clinics:
        items:
          $ref: '#/definitions/clinic.RqUpdateClinic'
        type: array
    required:
    - clinics
    type: object
  clinic.RqClinicAddress:
    properties:
      address:
        type: string
      city:
        type: string
      is_primary:
        type: boolean
      postcode:
        type: string
      state:
        type: string
    type: object
  clinic.RqClinicContact:
    properties:
      contact_type:
        enum:
        - PHONE
        - EMAIL
        - WEBSITE
        - FAX
        type: string
      is_primary:
        type: boolean
      label:
        type: string
      value:
        type: string
    required:
    - contact_type
    - value
    type: object
  clinic.RqCreateClinic:
    properties:
      abn:
        type: string
      addresses:
        items:
          $ref: '#/definitions/clinic.RqClinicAddress'
        type: array
      contacts:
        items:
          $ref: '#/definitions/clinic.RqClinicContact'
        type: array
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      profile_picture:
        type: string
    required:
    - name
    type: object
  clinic.RqUpdateAddress:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      postcode:
        type: string
      state:
        type: string
    type: object
  clinic.RqUpdateClinic:
    properties:
      abn:
        type: string
      addresses:
        items:
          $ref: '#/definitions/clinic.RqUpdateAddress'
        type: array
      contacts:
        items:
          $ref: '#/definitions/clinic.RqUpdateContact'
        type: array
      description:
        type: string
      financial_year_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      lock_date:
        type: string
      name:
        type: string
      profile_picture:
        type: string
    type: object
  clinic.RqUpdateContact:
    properties:
      contact_type:
        enum:
        - PHONE
        - EMAIL
        - WEBSITE
        - FAX
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      label:
        type: string
      value:
        type: string
    type: object
  clinic.RsClinic:
    properties:
      abn:
        type: string
      addresses:
        items:
          $ref: '#/definitions/clinic.RsClinicAddress'
        type: array
      contacts:
        items:
          $ref: '#/definitions/clinic.RsClinicContact'
        type: array
      created_at:
        type: string
      description:
        type: string
      financial_settings:
        $ref: '#/definitions/clinic.RsFinancialSettings'
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      practitioner_id:
        type: string
      profile_picture:
        type: string
      updated_at:
        type: string
    type: object
  clinic.RsClinicAddress:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      postcode:
        type: string
      state:
        type: string
    type: object
  clinic.RsClinicContact:
    properties:
      contact_type:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      label:
        type: string
      value:
        type: string
    type: object
  clinic.RsFinancialSettings:
    properties:
      financial_year_id:
        type: string
      id:
        type: string
      lock_date:
        type: string
    type: object
...
  response.RsError:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  util.RsList:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Backend API for acareca
  title: Backend API
  version: 1.0.0
paths:
...
  /clinic:
    get:
      parameters:
      - description: Filter by clinic name
        in: query
        name: name
        type: string
      - description: Filter by clinic ID
        in: query
        name: id
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Search across name, abn, description
        in: query
        name: search
        type: string
      - description: Sort field (name, is_active, created_at)
        in: query
        name: sort_by
        type: string
      - description: Sort direction (ASC, DESC)
        in: query
        name: order_by
        type: string
      - description: Page size (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/clinic.RsClinic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get all clinics for practitioner
      tags:
      - clinic
    post:
      consumes:
      - application/json
      parameters:
      - description: Clinic Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clinic.RqCreateClinic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clinic.RsClinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Create a new clinic
      tags:
      - clinic
  /clinic/{id}:
    delete:
      parameters:
      - description: Clinic UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Delete a clinic
      tags:
      - clinic
    get:
      parameters:
      - description: Clinic UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clinic.RsClinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get clinic by ID
      tags:
      - clinic
    put:
      consumes:
      - application/json
      parameters:
      - description: Clinic UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Clinic Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clinic.RqUpdateClinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clinic.RsClinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update clinic details
      tags:
      - clinic
  /clinic/bulk-delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bulk Delete Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clinic.RqBulkDeleteClinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Bulk delete clinics
      tags:
      - clinic
  /clinic/bulk-update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bulk Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clinic.RqBulkUpdateClinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Bulk update clinics
      tags:
      - clinic
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    description: Type "Bearer <your_token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"